# PaperLess.Frontend/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/api/correspondents" : {
      "get" : {
        "operationId" : "GetCorrespondents",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Correspondents" ]
      },
      "post" : {
        "operationId" : "CreateCorrespondent",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCorrespondent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Correspondents" ]
      }
    },
    "/api/correspondents/{id}" : {
      "delete" : {
        "operationId" : "DeleteCorrespondent",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Correspondents" ]
      },
      "put" : {
        "operationId" : "UpdateCorrespondent",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Correspondent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Correspondents" ]
      }
    },
...
    "/api/document_types" : {
      "get" : {
        "operationId" : "GetDocumentTypes",
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      },
      "post" : {
        "operationId" : "CreateDocumentType",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewDocumentType"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      }
    },
    "/api/document_types/{id}" : {
      "delete" : {
        "operationId" : "DeleteDocumentType",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      },
      "put" : {
        "operationId" : "UpdateDocumentType",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentType"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "DocumentTypes" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Correspondent" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 6,
          "document_count" : 1,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "last_correspondence" : "2000-01-23T04:56:07.000+00:00",
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "slug" : {
            "nullable" : true,
            "type" : "string"
          },
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "last_correspondence" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DocumentType" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 6,
          "document_count" : 1,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "id" : 0,
          "slug" : "slug"
        },
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "slug" : {
            "nullable" : true,
            "type" : "string"
          },
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "NewCorrespondent" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 0,
          "document_count" : 6,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match",
          "last_correspondence" : "2000-01-23T04:56:07.000+00:00"
        },
        "properties" : {
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "document_count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "last_correspondence" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "NewDocumentType" : {
        "additionalProperties" : false,
        "example" : {
          "matching_algorithm" : 0,
          "is_insensitive" : true,
          "name" : "name",
          "match" : "match"
        },
        "properties" : {
          "name" : {
            "nullable" : true,
            "type" : "string"
          },
          "match" : {
            "nullable" : true,
            "type" : "string"
          },
          "matching_algorithm" : {
            "format" : "int64",
            "type" : "integer"
          },
          "is_insensitive" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
    }
  }
}